# Boxes record for smaller Value enum.