--- a/crates/nu-command/src/conversions/into/record.rs
+++ b/crates/nu-command/src/conversions/into/record.rs
@@ -135,7 +135,7 @@ fn into_record(
                 .collect(),
             span,
         ),
-        Value::Record { val, .. } => Value::record(val, span),
+        Value::Record { val, .. } => Value::record(*val, span),
         Value::Error { .. } => input,
         other => Value::error(
             ShellError::TypeMismatch {