--- a/crates/nu-command/src/formats/to/nuon.rs
+++ b/crates/nu-command/src/formats/to/nuon.rs
@@ -252,7 +252,7 @@ pub fn value_to_string(
         )),
         Value::Record { val, .. } => {
             let mut collection = vec![];
-            for (col, val) in val {
+            for (col, val) in &**val {
                 collection.push(if needs_quotes(col) {
                     format!(
                         "{idt_po}\"{}\": {}",