@FilipAndersson245 2024-03-20
I ran the benchmark suit and saw no noticeable slowdown by the indirection of boxing records.