@sholderbach 2024-03-20
Mhh we currently don't have any benchmarks for heavy cellpath access or records in general. Probably a gap worth filling.

Personally I would like to see the boxing route compared to a move of `Record` to a `Vec<(String, Value)>` which would also knock the size down to the same ceiling.
(tradeoffs all around: small vs big column counts, pure lookup, mutation)