@kubouch 2024-03-21
Since this is a performance optimization, it would be great to have a benchmark demonstrating some performance benefit.