@devyn 2024-03-22
I think if this isn't a big performance regression and it improves memory usage, that would also be a win