@FilipAndersson245 2024-03-22
I added a small tests creation time, and access time with a simple record of 1-1000 columns
there seem to be no noticeable difference between the two

# Before
```
Timer precision: 10 ns
Timer precision: 10 ns
benchmarks            fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ record                           │               │               │               │         │
   ├─ access                        │               │               │               │         │
   │  ├─ 1            14.65 µs      │ 92.82 µs      │ 15 µs         │ 16.63 µs      │ 100     │ 100
   │  ├─ 10           15.77 µs      │ 34.09 µs      │ 16.14 µs      │ 17.2 µs       │ 100     │ 100
   │  ├─ 100          23.94 µs      │ 239.5 µs      │ 24.54 µs      │ 28 µs         │ 100     │ 100
   │  ╰─ 1000         99.09 µs      │ 206.9 µs      │ 102.5 µs      │ 106.2 µs      │ 100     │ 100
   ╰─ create                        │               │               │               │         │
      ├─ 1            21.12 µs      │ 51.25 µs      │ 21.45 µs      │ 22.25 µs      │ 100     │ 100
      ├─ 10           29.98 µs      │ 60.01 µs      │ 30.7 µs       │ 32.42 µs      │ 100     │ 100
      ├─ 100          130 µs        │ 293.7 µs      │ 132.5 µs      │ 143.4 µs      │ 100     │ 100
      ╰─ 1000         1.296 ms      │ 1.844 ms      │ 1.353 ms      │ 1.382 ms      │ 100     │ 100
```

# After
```
Timer precision: 10 ns
benchmarks            fastest       │ slowest       │ median        │ mean          │ samples │ iters
╰─ record                           │               │               │               │         │
   ├─ access                        │               │               │               │         │
   │  ├─ 1            16.81 µs      │ 30.32 µs      │ 17.19 µs      │ 17.64 µs      │ 100     │ 100
   │  ├─ 10           17.69 µs      │ 31.13 µs      │ 18.11 µs      │ 18.4 µs       │ 100     │ 100
   │  ├─ 100          24.96 µs      │ 41.66 µs      │ 25.41 µs      │ 26.26 µs      │ 100     │ 100
   │  ╰─ 1000         88.65 µs      │ 192.7 µs      │ 105.8 µs      │ 107.6 µs      │ 100     │ 100
   ╰─ create                        │               │               │               │         │
      ├─ 1            20.93 µs      │ 41.1 µs       │ 21.37 µs      │ 21.86 µs      │ 100     │ 100
      ├─ 10           30.25 µs      │ 58.25 µs      │ 30.79 µs      │ 31.64 µs      │ 100     │ 100
      ├─ 100          134.7 µs      │ 219.3 µs      │ 139.3 µs      │ 141 µs        │ 100     │ 100
      ╰─ 1000         1.206 ms      │ 1.448 ms      │ 1.315 ms      │ 1.301 ms      │ 100     │ 100
```